% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_activity_type_pod_measure_options.R
\name{get_pod_lookup}
\alias{get_pod_lookup}
\title{This reshapes the output of get_activity_type_pod_measure_options to
match the format required for reskit functions.
Ideally we shouldn't need this in future (issue #406)}
\usage{
get_pod_lookup()
}
\description{
This reshapes the output of get_activity_type_pod_measure_options to
match the format required for reskit functions.
Ideally we shouldn't need this in future (issue #406)
}
